refactor: create themes type file
