+++ src/components/icon/icon.style.ts
@@ -4,7 +4,8 @@ import { shade } from "polished";
 
 import { margin } from "styled-system";
 
-import baseTheme, { ThemeObject } from "../../style/themes/base";
+import baseTheme from "../../style/themes/base";
+import { ThemeObject } from "../../style/themes/theme.types";
 import addFocusStyling from "../../style/utils/add-focus-styling";
 import styledColor from "../../style/utils/color";
 import getColorValue from "../../style/utils/get-color-value";
