+++ src/components/switch/switch.style.ts
@@ -8,7 +8,8 @@ import FieldHelpStyle from "../../__internal__/field-help/field-help.style";
 import { FieldLineStyle } from "../../__internal__/form-field/form-field.style";
 import { StyledLabelContainer } from "../../__internal__/label/label.style";
 import StyledValidationIcon from "../../__internal__/validations/validation-icon.style";
-import baseTheme, { ThemeObject } from "../../style/themes/base";
+import baseTheme from "../../style/themes/base";
+import type { ThemeObject } from "../../style/themes/theme.types";
 import addFocusStyling from "../../style/utils/add-focus-styling";
 
 import { SwitchProps } from "./switch.component";
