+++ src/style/design-tokens/generate-css-variables.util.ts
@@ -1,4 +1,4 @@
-import { ThemeObject } from "../themes/base";
+import type { ThemeObject } from "../themes/theme.types";
 
 /**
  *
