+++ src/style/themes/base/index.ts
@@ -1,9 +1,10 @@
-import configureBase, { BasePalette, ThemeObject } from "./base-theme.config";
+import configureBase from "./base-theme.config";
 import baseColors from "../../color-config";
 import atOpacity from "../../utils/at-opacity";
 import { mergeDeep } from "../../utils/merge-deep";
 import generatePalette from "../../palette";
 import addHexSymbols from "../../utils/add-hex-symbols";
+import type { BasePalette, ThemeObject } from "../theme.types";
 
 const colors = generatePalette(baseColors);
 const colorsWithHex = addHexSymbols(baseColors);
@@ -31,5 +32,3 @@ export const mergeWithBase = (
     palette,
   };
 };
-
-export type { ThemeObject };
