+++ src/style/themes/index.ts
@@ -4,4 +4,4 @@ export { sageTheme };
 export { default as baseTheme } from "./base";
 export { default as noTheme } from "./none";
 
-export type { ThemeObject } from "./base";
+export type { ThemeObject } from "./theme.types";
