+++ src/style/utils/color.ts
@@ -1,7 +1,6 @@
-// eslint-disable-next-line no-restricted-imports
 import { color as styledColor } from "styled-system";
 import tokens from "@sage/design-tokens/js/base/common";
-import { ThemeObject } from "../themes/base";
+import type { ThemeObject } from "../themes/theme.types";
 
 /*
  * styled-system/color allows users to use a color from the theme, from the `colors` object.
